certificate-service/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.4.2
pydantic-settings==2.0.3
python-jose[cryptography]==3.3.0
httpx==0.25.1
python-dotenv==1.0.0
jinja2==3.1.2
